projecte-taller/api/requirements.txt
Flask
mysql-connector-python
flask_cors
